Break up definition service
