+++ language-service/language_service/service/definition/__init__.py
@@ -2,12 +2,12 @@
 Where you put definition business logic.
 We combine different data sources as available to deliver the best definitions.
 """
-from language_service.dto import ChineseDefinition
-from language_service.service.definition.cedict import CEDICT
-from language_service.service.definition.wiktionary import Wiktionary
+from multiprocessing.context import TimeoutError
+from multiprocessing.dummy import Pool
 
-cedict = CEDICT()
-wiktionary = Wiktionary()
+from language_service.service.definition.language.chinese import get_chinese_definitions
+from language_service.service.definition.language.english import get_english_definitions
+from language_service.service.definition.language.spanish import get_spanish_definitions
 
 
 def get_definitions(language, word):
@@ -15,50 +15,44 @@ def get_definitions(language, word):
         return get_chinese_definitions(word)
 
     elif language == "ENGLISH":
-        return wiktionary.get_definitions(language, word)
+        return get_english_definitions(word)
 
     elif language == "SPANISH":
-        return wiktionary.get_definitions(language, word)
+        return get_spanish_definitions(word)
 
     else:
         raise NotImplementedError("Unknown language requested: %s")
 
 
-def get_chinese_definitions(word):
-    wiktionary_definitions = wiktionary.get_definitions("CHINESE", word)
-    cedict_definition = cedict.get_definitions(word)
+def fetch_definitions(language, word):
+    """
+    Helper method to handle getting and deserializing a single definition
+    Can't be a lambda because it must be picked to work with the thread pool
+    """
+    definitions = get_definitions(language, word)
+    return word, definitions
 
-    if wiktionary_definitions is not None and cedict_definition is not None:
-        print("CHINESE - Found wiktionary and cedict definitions for %s" % word)
+
+def get_definitions_for_group(language, words):
+    with Pool(10) as pool:
+        # First we trigger the lookups in parallel here
+        requests = [
+            pool.apply_async(fetch_definitions, args=(language, word)) for word in words
+        ]
+
+        # Then we get the results
         definitions = []
-        print(cedict_definition)
-        for definition in wiktionary_definitions:
-            # The CEDICT definitions are more focused than wiktionary so we should prefer them.
-            subdefinitions = (
-                cedict_definition.subdefinitions
-                if cedict_definition.subdefinitions is not None
-                else definition.subdefinitions
-            )
-
-            improved_definition = ChineseDefinition(
-                subdefinitions=subdefinitions,
-                tag=definition.tag,
-                examples=definition.examples,
-                pinyin=cedict_definition.pinyin,
-                simplified=cedict_definition.simplified,
-                traditional=cedict_definition.traditional,
-            )
-            definitions.append(improved_definition)
+        for result in requests:
+            try:
+                # And then either the result is ready, or we time out.
+                word, word_definitions = result.get(5)
+                if word_definitions is not None:
+                    print("Got definitions in %s for %s" % (language, word))
+                    definitions.append((word, word_definitions))
+                else:
+                    print("No definition in %s found for %s" % (language, word))
+                    definitions.append((word, None))
+            except TimeoutError:
+                print("Definition lookup timed out")
+                definitions.append((word, None))
         return definitions
-
-    elif wiktionary_definitions is not None and cedict_definition is None:
-        print("CHINESE - Only found wiktionary definition for %s" % word)
-        return wiktionary_definitions
-
-    elif wiktionary_definitions is None and cedict_definition is not None:
-        print("CHINESE - Only found cedict definition for %s" % word)
-        return [cedict_definition]
-
-    else:
-        print("CHINESE - No definition found for %s" % word)
-        return None
