+++ language-service/language_service/service/definition/language/spanish.py
@@ -1,0 +1,7 @@
+from language_service.service.definition.sources.wiktionary import Wiktionary
+
+wiktionary = Wiktionary()
+
+
+def get_spanish_definitions(word):
+    return wiktionary.get_definitions("SPANISH", word)
